+++ packages/core/global.d.ts
@@ -1,5 +1,14 @@
 declare module 'whyframe:app' {
-  export const createApp: (el: HTMLElement) => {
+  interface CreateAppOptions {}
+
+  /**
+   * @param el The element to mount the app to
+   * @param opts Extra options added by framework plugins
+   */
+  export const createApp: (
+    el: HTMLElement,
+    opts?: CreateAppOptions
+  ) => {
     destroy: () => void
   }
 }