+++ packages/core/src/plugins/core.js
@@ -67,13 +67,13 @@ export function corePlugin() {
 }
 
 const devCode = `\
-export async function createApp(el) {
+export async function createApp(el, opts) {
   const iframe = window.frameElement
   if (!iframe) throw new Error('[whyframe] page is not within an iframe')
   const url = iframe.dataset.whyId
   if (!url) throw new Error('[whyframe] iframe does not have an id')
   const data = await import(/* @vite-ignore */ url)
-  const result = await data.createApp(el)
+  const result = await data.createApp(el, opts)
   return result
 }
 if (typeof window !== 'undefined' && window.frameElement) {
@@ -88,7 +88,7 @@ if (typeof window !== 'undefined' && window.frameElement) {
 
 const buildCode = `\
 import hashToImportMap from 'whyframe:build-data'
-export async function createApp(el) {
+export async function createApp(el, opts) {
   const iframe = window.frameElement
   if (!iframe) throw new Error('[whyframe] page is not within an iframe')
   const hash = iframe.dataset.whyId
@@ -96,6 +96,6 @@ export async function createApp(el) {
   const importApp = hashToImportMap[hash]
   if (!importApp) throw new Error('[whyframe] no app found')
   const data = await importApp()
-  const result = await data.createApp(el)
+  const result = await data.createApp(el, opts)
   return result
 }`